improv: core rewrite
