+++ src/app/core/system_status.rs
@@ -18,20 +18,18 @@ impl Deployment {
     pub async fn get_all() -> Result<Vec<Deployment>, anywho::Error> {
         use tokio::process::Command;
 
-        // try direct rpm-ostree if not in distrobox
-        let output = match Command::new("rpm-ostree").args(["status"]).output().await {
+        let output = if super::is_running_in_distrobox() {
+            Command::new("distrobox-host-exec")
+                .args(["rpm-ostree", "status"])
+                .output()
+                .await
+        } else {
+            Command::new("rpm-ostree").args(["status"]).output().await
+        };
+
+        let output = match output {
             Ok(output) => output,
-            Err(_) => {
-                // fallback: use distrobox-host-exec to run commands on the host system from within distrobox
-                match Command::new("distrobox-host-exec")
-                    .args(["rpm-ostree", "status"])
-                    .output()
-                    .await
-                {
-                    Ok(output) => output,
-                    Err(_) => return Err(anywho!("Error fetching System Status")),
-                }
-            }
+            Err(err) => return Err(anywho!("Error fetching System Status: {}", err)),
         };
 
         let stdout = String::from_utf8_lossy(&output.stdout);
@@ -143,40 +141,38 @@ impl Deployment {
     pub async fn pin_deployment(deployment_index: i32) -> Result<(), anywho::Error> {
         use tokio::process::Command;
 
-        // pkexec first
-        let output = Command::new("pkexec")
-            .args(["ostree", "admin", "pin", &deployment_index.to_string()])
-            .output()
-            .await?;
-
-        if output.status.success() {
-            return Ok(());
-        } else if output.status == ExitStatus::from_raw(32256) {
-            return Err(anywho!("Permision denied"));
+        let output = if super::is_running_in_distrobox() {
+            Command::new("distrobox-host-exec")
+                .args([
+                    "pkexec",
+                    "ostree",
+                    "admin",
+                    "pin",
+                    &deployment_index.to_string(),
+                ])
+                .output()
+                .await
+        } else {
+            Command::new("pkexec")
+                .args(["ostree", "admin", "pin", &deployment_index.to_string()])
+                .output()
+                .await
         };
-
-        // fallback to distrobox-host-exec
-        let output = Command::new("distrobox-host-exec")
-            .args([
-                "pkexec",
-                "ostree",
-                "admin",
-                "pin",
-                &deployment_index.to_string(),
-            ])
-            .output()
-            .await;
 
         match output {
             Ok(output) => {
                 if output.status.success() {
                     Ok(())
                 } else {
+                    if output.status == ExitStatus::from_raw(32256) {
+                        return Err(anywho!("Permision denied"));
+                    }
+
                     let stderr = String::from_utf8_lossy(&output.stderr);
                     let stdout = String::from_utf8_lossy(&output.stdout);
 
                     Err(anywho!(
-                        "distrobox-host-exec failed: {}",
+                        "{}",
                         if !stderr.is_empty() {
                             stderr.trim()
                         } else {
@@ -185,61 +181,52 @@ impl Deployment {
                     ))
                 }
             }
-            Err(err) => {
-                if err.kind() == tokio::io::ErrorKind::NotFound {
-                    return Err(anywho!(
-                        "Command not found, are you using a RPM OSTree System?"
-                    ));
-                }
-                Err(anywho!("Unknown error"))
-            }
+            Err(err) => Err(anywho!("{}", err)),
         }
     }
 
     pub async fn unpin_deployment(deployment_index: i32) -> Result<(), anywho::Error> {
         use tokio::process::Command;
 
-        // pkexec first
-        let output = Command::new("pkexec")
-            .args([
-                "ostree",
-                "admin",
-                "pin",
-                "--unpin",
-                &deployment_index.to_string(),
-            ])
-            .output()
-            .await?;
-
-        if output.status.success() {
-            return Ok(());
-        } else if output.status == ExitStatus::from_raw(32256) {
-            return Err(anywho!("Permision denied"));
+        let output = if super::is_running_in_distrobox() {
+            Command::new("distrobox-host-exec")
+                .args([
+                    "pkexec",
+                    "ostree",
+                    "admin",
+                    "pin",
+                    "--unpin",
+                    &deployment_index.to_string(),
+                ])
+                .output()
+                .await
+        } else {
+            Command::new("pkexec")
+                .args([
+                    "ostree",
+                    "admin",
+                    "pin",
+                    "--unpin",
+                    &deployment_index.to_string(),
+                ])
+                .output()
+                .await
         };
-
-        // fallback to distrobox-host-exec
-        let output = Command::new("distrobox-host-exec")
-            .args([
-                "pkexec",
-                "ostree",
-                "admin",
-                "pin",
-                "--unpin",
-                &deployment_index.to_string(),
-            ])
-            .output()
-            .await;
 
         match output {
             Ok(output) => {
                 if output.status.success() {
                     Ok(())
                 } else {
+                    if output.status == ExitStatus::from_raw(32256) {
+                        return Err(anywho!("Permision denied"));
+                    }
+
                     let stderr = String::from_utf8_lossy(&output.stderr);
                     let stdout = String::from_utf8_lossy(&output.stdout);
 
                     Err(anywho!(
-                        "distrobox-host-exec failed: {}",
+                        "{}",
                         if !stderr.is_empty() {
                             stderr.trim()
                         } else {
@@ -248,14 +235,7 @@ impl Deployment {
                     ))
                 }
             }
-            Err(err) => {
-                if err.kind() == tokio::io::ErrorKind::NotFound {
-                    return Err(anywho!(
-                        "Command not found, are you using a RPM OSTree System?"
-                    ));
-                }
-                Err(anywho!("Unknown error"))
-            }
+            Err(err) => Err(anywho!("{}", err)),
         }
     }
 }
