+++ src/app/screen/system_status.rs
@@ -5,7 +5,7 @@ use iced::widget::{Rule, Space, button, column, container, row, scrollable, text
 use iced::{Alignment, Element, Length, Subscription, Task};
 
 use crate::app::core::system_status::Deployment;
-use crate::app::core::utils::{reboot, reboot_pending};
+use crate::app::core::{reboot, reboot_pending};
 use crate::app::style::{icon_button_style, icon_svg_style, primary_button_style};
 use crate::app::widgets::spinners::circular::Circular;
 use crate::app::widgets::spinners::easing;
